Update alternative locations unit tests for ARM64
Similar to the preceding commit but for unit-style tests rather
than integration-style tests, this updates the unit tests for
Windows-specific `git` executable alternative locations to include
`clangarm64/bin` directories when looking under the 64-bit program
files directory.

(The tests that are updated here are those that are true unit
tests, because they examine the `locations_under_program_files()`
helper function that takes a dependency-injected `var_os()`-like
function, rather than examining `ALTERNATIVE_LOCATIONS` itself.)

As in the integration-style tests updated in the preceding commit,
these tests will now fail until the logic for enumerating
alternative locations to check for the `git` executable is fixed to
include a location for ARM64, under `clangarm64` (as well as still
checking for a x86_64 build, under `mingw64`).