Test the `EXEPATH` optimization of `system_prefix()`
This tests the `EXEPATH` strategy that `system_prefix()` attempts
on Windows (before falling back to the `git --exec-path` strategy).
Specifically, this tests the `system_prefix_from_exepath_var()`
helper that was recently introduced to facilitate testing.

Some of these new tests don't pass yet, because they test two
scenarios that are not yet covered:

- More than one of `mingw32`, `mingw64`, and `clangarm64` exist
  under the directory named by the value of `EXEPATH`. In this
  situation, the `EXEPATH` optimization should not be used.

- `clangarm64` exists under the directory named by the value of
  `EXEPATH` (and neither `mingw32` nor `mingw64` exist under it).
  In this situation, the `clangarm64` subdirectory should be used.

Although the `EXEPATH` optimization is only ever used on Windows,
nothing in its test or implementation precludes testing elsewhere,
so the tests build and run on Unix-like systems too.