Rename `PlatformArchitecture` helper `PlatformBitness`
The `PlatformArchitecture` test helper struct holds information
about whether we are 64-bit or 32-bit and whether the system we are
running on is 64-bit or 32-bit. It doesn't hold information about
architecture beyond that. The name `PlatformArchitecture` made
sense in #1456 when this fully captured the relevant architecture
distinctions. But now, although `gix-path` doesn't work differently
between x86_64 and ARM64 systems, the distinction betwee x86_64 and
ARM64 is important in the design.

This renames `PlatformArchitecture` to `PlatformBitness`, while
names that refer to architecture in a way not completely reducible
to a 32-bit vs. 64-bit distinction are not renamed "bitness."