Update alternative locations integration tests for user program files
This updates the integration-style tests for Windows-specific `git`
executable alternative locations to include locations under the
per-user program files directory. The user program files directory,
when present, is a subdirectory of the user's local application
data directory. (The sense in which these are "integration-style"
tests is noted in e9770a7 in #2115, where they were last updated.)

These tests will now fail until `ALTERNATIVE_LOCATIONS` is enhanced
to check in the user program files directory (as well as the global
program files directories that are currently checked).