Add a "SAFETY:" comment in `PlatformBitness::current()`
The `GetCurrentProcess` and `IsWow64Process` Windows API calls
(mediated through the `windows` crate) are pretty straightforward,
and this code is only in the test suite, not in production. Still,
it's probably best to have a "SAFETY:" comment where `unsafe` code
is used. This adds such a comment, covering those two calls.