governance: add Benjamin Tan and Scott Taylor as maintainers
Copying from Discord:

@ilyagr recently accepted a job offer from East River Source Control
(ersc.io). That means we will have two of the five maintainers from
that company (@thoughtpolice is the other one), which is more than the
1/3 that our policy allows
(https://github.com/jj-vcs/jj/blob/main/GOVERNANCE.md#single-company-influence). To
resolve it, we decided to add more maintainers. We're happy to
announce that both @bnjmnt4n and @scott2000 have agreed to join. Thank
you!