Switch CodeQL to advanced config to run on more PRs
Because the basic configuration doesn't run on PRs from forks.

So far this is just the advanced configuration workflow file
written automatically when enabling it through the GitHub interface
for doing so, with no customizations. This should already be
sufficient to let it run on PRs from forks, but the immediately
forthcoming commits shall apply some customizations.