Keep showing debug representation of paths in gitoxide-core
Where this was done already, in `bail!` and `anyhow!`.

A newly enforced clippy lint warns about this, making the good
point that the debug representation can change in the future. But
the display representation it recommends using is less suitable in
these places, because it can result in more ambiguous output due to
the absence of quoting and escaping.

(In some cases, it seems like the display representation could
slightly exacerbate CVE-2024-43785 / #1534, but using the debug
representation on these two macro calls is peripheral to the cases
of greatest concern, and this doesn't really mitigate that.)

The clippy lint being suppressed here in two specific places is:
https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_debug_formatting