lib: add `ConflictLabels` type to store conflict labels
Adding this as a separate type will help maintain the invariant that
resolved merges cannot have labels.

I think that storing separate conflict labels for each term of the
conflict is the best approach for a couple reasons. Mainly, I think it
integrates well with the existing conflict algebra. For instance, a diff
of (A - B) and a diff of (B - C) can be easily combined to create a new
diff of (A - C), and if we associate a label with each term, then the
labels will also naturally be carried over as well. Also, I think it
would be simpler to implement than other approaches (such as storing
labels for diffs instead of terms), since conflict labels can re-use
existing logic from `Merge<T>`.