Improve support for sparse cache.
With the addition of a notes backed cache we will have the
possibility of avoiding a "cold start" by transferring just
the notes to another repo. To keep the amount of extra data
small the notes cache is kept sparse: Not all commits have
and entry and also in shards by sequence number, as old
entries are unlikely to be relevant.
The old traversal logic did not perform very well with sparse
cache. Especially "find_known" proved to be a bottleneck.
So the traversal is revised now to work better in the sparse
cache case.

Change: cache-sparse-shards