lib: add find_dominator_value_with_value_cache method
This commit adds `find_dominator_value_with_value_cache`. This API is more
convenient than the simpler `find_dominator_value` because it allows the caller
to use the ValueCache later on to get the cached values of nodes, and to obtain
all nodes that have produced a given cached value.

I had to rework the ValueCache API a bit. The idea is that now it is perfectly
safe for the caller to pass the ValueCache in, even if the caller had already
invoked some methods on the value cache object, that is, even if there are
already entries in the cache. The cardinality comparisons are gone.

Also, find_dominator_value and find_dominator_value_with_value_cache now take
the final_nodes as IntoIterator<Item = &'a N>. This is a bit more flexible.