working_copy: determine staleness by comparing op id instead of tree id
Our detection of stale working copy is based on the tree id since a
long time ago, at least since Feb 2022 (e098c0193566) depending on how
you count. Since Sep 2022 (443e73f346be), we keep the last operation
recorded in the working copy up to date. However, we don't update it
when the tree id matches. That's inconsistent, so I think we should
always keep it up to date. This patch fixes that.

Thanks to @kevincliao for spotting this. We noticed this at Google
because it meant that we sometimes didn't notice the new operation id
in our distributed file system, which led to the machine creating
divergent operations.  (The machine is supposed to detect operations
recorded in the operation log but this is sometimes flaky for
unrelated reasons.)