ci: macos-x86_64 should actually build for x86_64-apple-darwin
The `macos-15` runner is actually an ARM64 machine -- see
https://github.com/actions/runner-images for current runner image
meanings, but as of this commit, macos-15 was listed as this:

    Image 	YAML Label 	Included Software
    macOS 15 Arm64 	macos-latest, macos-15, or macos-15-xlarge 	macOS-15-arm64

This fixes things such that there is an _actual_ CI build of
x86_64-apple-darwin.

The release.yml and binaries.yml workflows are not affected, as they
manually specify targets, rather than taking the "autodetected" host
target (which is what was happening here and is wrong for building
x86_64 binaries on aarch64 hardware via Rosetta).