working_copy: existing submodule directory on checkout is ok
Fixes #8065.

This commit fixes the problem of jj marking a git submodule as a deleted path,
after the user moves (edit/new) to a commit where the submodule did not exist
and then back to a commit where it existed. Failure to remove a submodule's
directory is no longer considered a conflict - if it was not going to be
replaced by another file.


The following transition cases during checkout have been modified:

disk:              directory
before tree value: <don't care>
after tree value:  submodule

Problem: update() would try to remove the directory (remove_old_file()) or write
in its place (can_create_new_file()) and fail, indicating a conflict to the
user.

Fix: Consider a submodule checkout to be successful even if a directory exists
at the submodule's path. In fact, assume that the directory is the submodule's
directory (possibly left over when moving between commits in the past). This is
in line with Git's behavior.


disk:              directory
before tree value: submodule
after tree value:  absent

Problem: update() would try to remove the directory (remove_old_file()) and
fail, indicating a conflict to the user.

Fix: Consider a failure to remove a submodule directory to not be a failure, if
the commit does not need to put another file in the submodule's place.


In the general case, failing to remove a file, that should not exist after the
checkout is complete, is a failure. However, for a (previous) submodule
directory, there is no safe way to do the remove, as the directory could contain
non-commited or local-only changes. Therefore, in the case of the commit not
wanting to place a new file at the same path, consider the checkout of the
removed submodule successful. This is in line with Git's behavior.